Refactored processor class
